Added check whether directory or file exists before deleting

+++ windows-client/DbUtils/MiscDBUtil.cs
@@ -270,8 +270,10 @@ namespace windows_client.DbUtils
             string attachmentFileBytes = HikeConstants.FILES_BYTE_LOCATION + "/" + msisdn + "/" + Convert.ToString(messageId);
             using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
             {
-                store.DeleteFile(attachmentObjectPath);
-                store.DeleteFile(attachmentFileBytes);
+                if (store.FileExists(attachmentObjectPath))
+                    store.DeleteFile(attachmentObjectPath);
+                if (store.FileExists(attachmentFileBytes))
+                    store.DeleteFile(attachmentFileBytes);
             }
         }
 
@@ -285,10 +287,13 @@ namespace windows_client.DbUtils
             {
                 foreach(string attachmentPath in attachmentPaths)
                 {
-                    string[] fileNames = store.GetFileNames(attachmentPath + "/*");
-                    foreach (string fileName in fileNames)
+                    if (store.DirectoryExists(attachmentPath))
                     {
-                        store.DeleteFile(attachmentPath + "/" + fileName);
+                        string[] fileNames = store.GetFileNames(attachmentPath + "/*");
+                        foreach (string fileName in fileNames)
+                        {
+                            store.DeleteFile(attachmentPath + "/" + fileName);
+                        }
                     }
                 }
             }
